Move roulette mini-game into pari_xp package and add entry point
